more components fixing for one comment page.
