+++ src/components/Comments/Comments.css
@@ -19,7 +19,9 @@ ul.commentsContainer {
 
 li.comment {
   /* background-color: aqua; */
-  list-style-position: inside;
+  /* list-style-position: inside; */
+  list-style: none;
+  padding-left: 0;
   display: flex;
   flex-direction: column;
   align-items: center;
