+++ src/components/Comments/Comments.js
@@ -1,6 +1,7 @@
 import { CreateComment } from '../../pages/CreateComment/CreateComment'
 import { deleteCommentDEL } from '../../utils/functions/deleteCommentDEL'
 import { BasicBox } from '../BasicBox/BasicBox'
+import { singleComment } from '../singleComment/singleComment'
 import './Comments.css'
 
 export const commentsList = (arrayOfComments, title) => {
@@ -18,41 +19,7 @@ export const commentsList = (arrayOfComments, title) => {
     for (const comment of arrayOfComments) {
       const listItem = document.createElement('li')
       listItem.classList.add('comment')
-
-      listItem.append(BasicBox(`<p>UniqueID: ${comment._id}</p>`, 'divID'))
-      listItem.append(BasicBox(`<p>${comment.description}</p>`, 'description'))
-      const createDateTime = new Date(comment.createdAt)
-      const updateDateTime = new Date(comment.updatedAt)
-      const dateTimeContent = `  <p class="type-comment">Created ${createDateTime.toUTCString()}</p> <p class="type-comment">updated ${updateDateTime.toUTCString()}</p>`
-      listItem.append(BasicBox(dateTimeContent, 'dateTime'))
-      const statusContent = ` <div class="small-boxes"> <p class="type-comment">Type: ${
-        comment.typeComment
-      }</p> </div><div class="small-boxes"><p class="resolved">Complete: ${
-        comment.resolved ? '✅' : '❎'
-      } </div><div class="small-boxes"> <p class="who">Written by ${
-        comment.person.personName
-      }</p></div>`
-      listItem.append(BasicBox(statusContent, 'status'))
-
-      if (
-        comment.relatedComments.length > 0 &&
-        comment.relatedComments[0] !== ''
-      ) {
-        const divRelatedComments = document.createElement('div')
-        divRelatedComments.classList.add('comment-box')
-        divRelatedComments.innerHTML = `<p class="related"> Related to: ${comment.relatedComments}</p>`
-        listItem.append(divRelatedComments)
-      }
-      if (comment.img.length == 0) {
-        //do nothing
-      } else {
-        const figure = document.createElement('figure')
-        for (const img of comment.img) {
-          figure.innerHTML += `<img src="${img}"/>`
-        }
-        listItem.append(figure)
-        ul.append(listItem)
-      }
+      listItem.append(singleComment(comment))
 
       const divActions = document.createElement('div')
       divActions.classList.add('actions')
