--- src/components/singleComment/singleComment.jsx
@@ -1,11 +1,0 @@
-import './singleComment.css';
-
-const SingleComment = () => {
-  return (
-    <div className='singlecomment'>
-        <h1>SingleComment Component</h1>
-    </div>
-  );
-};
-
-export default SingleComment;
