+++ src/pages/EditOneComment/EditOneComment.js
@@ -1,0 +1,5 @@
+export const Home = async () => {
+  console.log('Got to my home site')
+  const div = createPage('One Comment')
+  div.innerHTML = `<h2>Edit Comment</h2>`
+}
