+++ src/utils/functions/typeCommentsGET.js
@@ -2,11 +2,11 @@ import { API } from '../API/API'
 
 export const typeCommentsGET = async (type) => {
   // e.preventDefault()
-  console.log('Getting my comments')
+  console.log('Getting comments')
   const token = JSON.parse(localStorage.getItem('token'))
   const id = JSON.parse(localStorage.getItem('person'))._id
-  console.log('this is the token', token)
-  console.log('this is the Id', id)
+  // console.log('this is the token', token)
+  // console.log('this is the Id', id)
   try {
     const data = await API({
       endpoint: `comments/type/${type}`,
@@ -16,7 +16,7 @@ export const typeCommentsGET = async (type) => {
         Authorization: `Bearer ${token}`
       }
     })
-    console.log('The data of my comments', data)
+    console.log('Got the the data comments', type)
     return data
   } catch (error) {
     console.error('Error fetching comments:', error)
